// src/app/models/user.model.ts
export class UserModel {
	uid? :string;
	id?: string;
  	name?: string;
	lastname?: string;
 	email?: string;
	tlf?: string;
	wapp?: string;
	address?: string;
	country?: string;
	city?: string;
	registration_date?: Date;
	lastUpdate?: Date;
	lastUpdateUser?: string;
	user_type?: string;
  	password?: string;
	profile_picture?: string;
	state?: string;
	status?: boolean;
}